requirements.txt
pydantic>=2.4,<3.0
pydantic-settings>=2.0,<3.0
pyyaml>=6.0,<7.0
pandas>=2.2,<3.0
sqlalchemy>=2.0,<3.0
mysql-connector-python>=9.0,<10.0
python-dotenv>=1.0,<2.0
